requirements.txt
fastapi
uvicorn
numpy
pillow
slowapi
httpx
python-multipart